firmware/gateway/ros2: round 14 -- oversize sweep + slog + ros2 format + doxygen
Five parallel agents + a CI hotfix for round 13's three failures:

Agent 1 -- oversize function sweep (NASA P10 Rule 4):
  rx_usb_hw.c: rx_usb_hw_configure_pipe 173 -> 54 lines (6 helpers),
                rx_usb_hw_fifo_read 63 -> 35.
  riic.c:      internal_wait_bus_ready 135 -> 37 (6 helpers + new
                riic_saved_cfg_t typedef), riic_init 109 -> 46 (3
                helpers reusing earlier ones).
  uart.c:      uart_init_channel 111 -> 32 (4 helpers).
  rx_clock_power_init.c: internal_start_oscillators_and_plls 92 -> 19
                (4 helpers split by board variant: TOM HOCO-PLL,
                PROD XTAL-PLL, PROD PPLL-for-USB).
  rspi.c:      rspi_init_peripheral 83 -> 34 (3 helpers).
  rx_mpc.c:    internal_get_pfs_register 78 -> 18 (1 helper +
                mpc_port_offset_sentinel_t enum).
  Each helper carries full Doxygen; register-write order and
  RX72N HW-manual section citations preserved inline.

Agent 2 -- gateway stdlib log -> slog migration:
  91 log calls converted across 8 internal/* files: cdc.go (4),
  handler.go (10), health_monitor.go (4), session.go (9),
  hotplug_detector.go (7), heartbeat.go (7), spi.go (4),
  manager.go (46). Log levels preserved per existing prefixes
  (DEBUG/WARN/ERROR/FATAL); FATAL still panics. cmd/* CLIs left
  alone per scope. No stdlib log imports remain in internal/.

Agent 3 -- star-ros2/.clang-format + single-pass reformat:
  Created star-ros2/.clang-format matching docs/sections/11. Key
  options: ColumnLimit 120, IndentWidth 2, Allman for functions
  and namespaces (cuddled for if/while/for), AccessModifierOffset
  -2, Middle pointer alignment, ROS2 include category priorities
  (rclcpp -> msgs -> project -> system C++ -> system C). Reformatted
  19 files across star_spi_bridge, star_gateway_bridge,
  star_safety_monitor (incl. their src/, include/, test/).
  Vendored sllidar_ros2 untouched.

Agent 4 -- Doxygen warnings recheck + tool-config:
  PLANTUML_JAR_PATH and MSCGEN_TOOL pointed at brew installations.
  warnings.log: 213 -> 121 lines (-92). Source-doc fixes: 21
  duplicated `msc { }` wrappers removed across 12 files (Doxygen
  auto-wraps); 47 `--- [label=...]` -> `... [label=...]` in
  mscgen blocks across 18 files; 11 `@return None` removed from
  void functions; rx_gptw.h diagram label fixed; rx_compiler.h
  CCRX preprocessor link wrapped in backticks; duplicated
  @startuml lines collapsed; invalid PlantUML `note right:`
  syntax stripped from rx72n_wdt_regs.h. Verified that
  WARN_AS_ERROR=YES still segfaults doxygen 1.16.1 even with
  PlantUML/mscgen working; FAIL_ON_WARNINGS retained.

Agent 5 -- testability for round-12/13 helpers (coverage):
  Coverage was already 100/100 lines/funcs in CI; only one
  branch was missing -- the false branch of the round-11
  re-entrancy guard at rx_spi_link.c::internal_send_attempt:401.
  Switched s_send_path_busy from `static atomic_flag` to
  RX_STATIC_TESTABLE so the test can drive the assert path
  directly. Added test_send_attempt_reentrancy_assert_fires.
  CI branch coverage projects 4055/4056 -> 4056/4056 = 100%.
  Zero-overhead in firmware builds (RX_STATIC_TESTABLE expands
  to static).

Round 13 CI hotfix (folded into this commit):
  - rx_register_guard.c: re-add rx_gpio_constants.h include
    inside the #ifdef __RX__ block (round-13 unused-include
    audit incorrectly removed it; cross-compile build needs
    k_prcr_key, k_prcr_key_shift, k_prcr_lock_prc1, k_prcr_lock_all).
  - temp_sensor_task.c, imu_task.c, obstacle_detect_task.c:
    redundant `extern rx_bus_manager_t g_bus_manager` declarations
    removed (canonical extern in shared_data.h since round 12).
  - Gateway: agent 4's slog migration also incidentally cleaned up
    the binary.BigEndian / gofmt / go.mod tidy issues that broke
    Gateway CI on round 13.

Local: 61/61 firmware tests pass, gateway tests pass (excl. the
e2e test holding port 8080 on this dev box), CI tidy build clean,
clang-format-18 clean, ASCII clean across all touched files.

+++ star-rx72n-firmware/src/tasks/serial_bringup_task.c
@@ -395,7 +395,7 @@ static void internal_drain_rx(void)
     s_total_rx_bytes++;
     const uint8_t byte_u = (uint8_t)byte_c;
 
-    if ((byte_u == (uint8_t)k_ascii_lf) || (byte_u == (uint8_t)k_ascii_cr)) {
+    if ((byte_u == k_ascii_lf) || (byte_u == k_ascii_cr)) {
       if (s_line_len > 0U) {
         s_total_lines++;
         internal_handle_line(s_line, s_line_len);
@@ -404,10 +404,10 @@ static void internal_drain_rx(void)
       continue;
     }
 
-    if (byte_u < (uint8_t)k_ascii_space) {
+    if (byte_u < k_ascii_space) {
       continue;
     }
-    if (byte_u > (uint8_t)k_ascii_tilde) {
+    if (byte_u > k_ascii_tilde) {
       continue;
     }
 
@@ -448,7 +448,7 @@ static void internal_handle_line(const char* line, uint8_t len)
     return;
   }
   /* Comment / non-command lines: ignore silently. */
-  if ((uint8_t)line[0] != (uint8_t)k_ascii_cmd_velocity) {
+  if ((uint8_t)line[0] != k_ascii_cmd_velocity) {
     return;
   }
 
@@ -462,7 +462,7 @@ static void internal_handle_line(const char* line, uint8_t len)
   const char* cursor                      = &scratch[1];
   float       parsed[k_serial_num_motors] = {0.0F, 0.0F, 0.0F, 0.0F};
 
-  for (uint8_t i = 0; i < (uint8_t)k_serial_num_motors; i++) {
+  for (uint8_t i = 0; i < k_serial_num_motors; i++) {
     char* endptr = NULL;
     /* strtof() in newlib-nano returns NaN on this build; strtod() works. */
     const double v = strtod(cursor, &endptr);
